MXS-4618: Use leint in create_ok_packet()
The code did not support length-encoded values that were not
representable with only one byte. For proper row count tracking, the
function needs to implement it correctly.

+++ include/maxscale/protocol/mariadb/mysql.hh
@@ -203,7 +203,7 @@ bool is_com_prepare(const GWBUF& buf);
  */
 bool is_com_query_or_prepare(const GWBUF& buf);
 
-GWBUF create_ok_packet(uint8_t sequence, uint8_t affected_rows);
+GWBUF create_ok_packet(uint8_t sequence, uint64_t affected_rows);
 
 GWBUF create_query(std::string_view query);
